+++ src/pages/products/components/FilterElement/FilterElement.js
@@ -32,7 +32,7 @@ class FilterElement extends Component {
         <UncontrolledButtonDropdown>
           <DropdownToggle
             caret color="default"
-            className="dropdown-toggle-split mr-xs"
+            className={`dropdown-toggle-split mr-xs ${s.back}`}
           >
             {currentOption}&nbsp;&nbsp;
           </DropdownToggle>